用户账号相关

+++ src/main/java/com/ngtesting/platform/model/TstVerifyCode.java
@@ -6,7 +6,7 @@ public class TstVerifyCode {
     private String code;
     private Date expireTime;
 
-    private Long refId;
+    private Integer refId;
 
     public String getCode() {
         return code;
@@ -24,11 +24,11 @@ public class TstVerifyCode {
         this.expireTime = expireTime;
     }
 
-    public Long getRefId() {
+    public Integer getRefId() {
         return refId;
     }
 
-    public void setRefId(Long refId) {
+    public void setRefId(Integer refId) {
         this.refId = refId;
     }
 }
